+++ chempy/tests/test_equilibria.py
@@ -13,8 +13,7 @@ from ..chemistry import (
 
 from ..equilibria import (
     equilibrium_quotient, equilibrium_residual, get_rc_interval,
-    solve_equilibrium, EqSystemBase, prodpow,
-    EqSystemLin, EqSystemLog
+    solve_equilibrium, prodpow, EqSystem, NumSysLin, NumSysLog
 )
 
 from .ammonical_cupric_solution import get_ammonical_cupric_eqsys
@@ -66,20 +65,20 @@ def test_solve_equilibrium_2():
     assert np.allclose(solution, c + stoich*fsolve(f, 0.1))
 
 
-def test_EqSystemBase():
+def test_EqSystem():
     a, b = sbstncs = Substance('a'), Substance('b')
     rxns = [Reaction({a: 1}, {b: 1})]
-    es = EqSystemBase(rxns, sbstncs)
+    es = EqSystem(rxns, sbstncs)
     assert es.net_stoichs().tolist() == [[-1, 1]]
 
 
 def _get_es1():
     a, b = sbstncs = Solute('a'), Solute('b')
     rxns = [Equilibrium({a: 1}, {b: 1}, 3)]
-    return EqSystemBase(rxns, sbstncs)
+    return EqSystem(rxns, sbstncs)
 
 
-def _get_es_water(EqSys=EqSystemBase):
+def _get_es_water(EqSys=EqSystem):
     H2O = Solute('H2O', charge=0, composition={1: 2, 8: 1})
     OHm = Solute('OH-', charge=-1, composition={1: 1, 8: 1})
     Hp = Solute('H+', charge=1, composition={1: 1})
@@ -88,7 +87,7 @@ def _get_es_water(EqSys=EqSystemBase):
     return EqSys([w_auto_p], [H2O, OHm, Hp])
 
 
-def test_EqSystemBase_1():
+def test_EqSystem_1():
     es = _get_es1()
     assert es.stoichs().tolist() == [[-1, 1]]
     assert es.eq_constants() == [3]
@@ -104,7 +103,7 @@ def test_Equilibria_arithmetics():
 
 def test_Equilibria_root():
     eqsys, c0 = get_ammonical_cupric_eqsys()
-    x, sol = eqsys.root(c0)
+    x, sol = eqsys.root(c0, NumSys=(NumSysLog,))
     assert sol.success
 
 
@@ -122,13 +121,12 @@ def test_Equilibria_root_simple():
     ammonia_protolysis = Equilibrium(
         {ammonium: 1}, {hydronium: 1, ammonia: 1}, 10**-9.26/55.5
     )
-    eqsys_log, eqsys_lin = [EqSys([water_auto_protolysis,
-                                   ammonia_protolysis], solutes)
-                            for EqSys in (EqSystemLog, EqSystemLin)]
+    eqsys = EqSystem([water_auto_protolysis, ammonia_protolysis], solutes)
+
     init_concs = collections.defaultdict(float, {water: 55.5, ammonia: 1e-3})
-    x, sol1 = eqsys_log.root(init_concs)
-    x, sol2 = eqsys_lin.root(init_concs, x)
-    ref = eqsys_lin.as_per_substance_array({
+    x, sol1 = eqsys.root(init_concs)
+    x, sol2 = eqsys.root(init_concs, x, NumSys=(NumSysLog,))
+    ref = eqsys.as_per_substance_array({
         water: 55.5,
         ammonia: 1e-3 - 6.2e-4,
         ammonium: 6.2e-4,
@@ -138,17 +136,17 @@ def test_Equilibria_root_simple():
     assert np.allclose(x, ref, rtol=0.02, atol=1e-16)
 
 
-def _get_NaCl(EqSys):
+def _get_NaCl():
     Na_p, Cl_m, NaCl = sbstncs = (
         Solute('Na+', 1, composition={11: 1}),
         Solute('Cl-', -1, composition={17: 1}),
         Solute('NaCl', composition={11: 1, 17: 1}, solid=True)
     )
     sp = Equilibrium({NaCl: 1}, {Na_p: 1, Cl_m: 1}, 4.0)
-    eqsys = EqSys([sp], sbstncs)
+    eqsys = EqSystem([sp], sbstncs)
     cases = [
-        [(.5, .5, .4), (.9, .9, 0)],
         [(0, 0, .1), (.1, .1, 0)],
+        [(.5, .5, .4), (.9, .9, 0)],
         [(1, 1, 1), (2, 2, 0)],
         [(0, 0, 2), (2, 2, 0)],
         [(0, 0, 3), (2, 2, 1)],
@@ -158,15 +156,14 @@ def _get_NaCl(EqSys):
     return eqsys, (Na_p, Cl_m, NaCl), cases
 
 
-def test_EqSystemLog():
-    pass
-
-
-@pytest.mark.parametrize('EqSys', [EqSystemLin, EqSystemLog])
-def test_solid(EqSys):
-    eqsys, species, cases = _get_NaCl(EqSys)
+@pytest.mark.parametrize('NumSys', [(NumSysLin,), (NumSysLog,)])
+def test_solid(NumSys):
+    eqsys, species, cases = _get_NaCl()
 
     for init, final in cases:
-        x, sol = eqsys.root(dict(zip(species, init)), method='lm')
+        print(init, final)
+        x, sol = eqsys.root(dict(zip(species, init)), method='lm',
+                            NumSys=NumSys)
+        print(sol)
         assert x is not None
         assert np.allclose(x, np.asarray(final))