+++ CHANGES.rst
@@ -1,8 +1,10 @@
 v0.2.0
 ======
+- EqSystemLog, EqSystemLin -> EqSystem, (NumSysLog, NumSysLin)
 - Support for solid phases in equilibria
 - Submodules for water properties moved to chempy.properties
 - Moved class ``Equilibrium`` from .equilibria to .chemistry
+- Renamed Reaction.params to Reaction.param
 
 v0.1
 ====